added more functionality

+++ main.cpp
@@ -6,7 +6,7 @@ using namespace std;
 int main() {
 	
 	WelcomePrompt();
-	double user_weight = GetUserWeight();
+	int user_weight = GetUserWeight();
 	int weight_loss_choice = GetUserLossOption();
 	CalculateUserCalories(weight_loss_choice, user_weight);
